# internal/controllers/accounts/x_gen_admin.go
// Code generated by core_gen; DO NOT EDIT.

package accounts

import (
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/griffnb/core/lib/log"
	"github.com/griffnb/core/lib/router/request"
	"github.com/griffnb/core/lib/router/response"
	"github.com/griffnb/core/lib/tools"
	"github.com/griffnb/core/lib/types"
	"github.com/griffnb/techboss-ai-go/internal/constants"
	"github.com/griffnb/techboss-ai-go/internal/models/account"
	"github.com/pkg/errors"
)

//	 Gets a list of Account entries
//
//		@Title			List Account
//
//		@Summary		List Account
//		@Description	List Account
//		@Tags			Account
//		@Accept			json
//		@Produce		json
//		@Param			q	query		string	false	"search by q"
//		@Param			limit	query		int		false	"limit results"		default(100)	minimum(1)	maximum(1000)
//		@Param			offset	query		int		false	"offset results"	default(0)	minimum(0)
//		@Param			order	query		string	false	"sort results e.g. 'created_at desc'"	default(created_at desc)
//		@Param			filters	query		string	false	"filters, see readme"
//		@Success		200	{object}	response.SuccessResponse{data=[]account.AccountJoined}
//		@Failure		400	{object}	response.ErrorResponse
//		@Failure		404	{object}	response.ErrorResponse
//		@Failure		500	{object}	response.ErrorResponse
//		@Router			/admin/account [get]
func adminIndex(_ http.ResponseWriter, req *http.Request) ([]*account.AccountJoined, int, error) {

	parameters := request.BuildIndexParams(req.Context(), req.URL.Query(), TABLE_NAME)

	if tools.Empty(parameters.Limit) {
		parameters.Limit = constants.SYSTEM_LIMIT
	}

	if !tools.Empty(req.URL.Query().Get("q")) {
		addSearch(parameters, req.URL.Query().Get("q"))
	}

	accountObjs, err := account.FindAllJoined(req.Context(), parameters)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.AdminBadRequestError[[]*account.AccountJoined](err)

	}

	return response.Success(accountObjs)

}

//	 Gets a  Account entry by ID
//
//		@Title			Get Account
//
//		@Summary		Get Account
//		@Description	Get Account
//		@Tags			Account
//		@Accept			json
//		@Produce		json
//		@Param			id	path		string	true	"Account ID"
//		@Success		200	{object}	response.SuccessResponse{data=account.AccountJoined}
//		@Failure		400	{object}	response.ErrorResponse
//		@Failure		404	{object}	response.ErrorResponse
//		@Failure		500	{object}	response.ErrorResponse
//		@Router			/admin/account/{id} [get]
func adminGet(_ http.ResponseWriter, req *http.Request) (*account.AccountJoined, int, error) {
	id := chi.URLParam(req, "id")

	accountObj, err := account.GetJoined(req.Context(), types.UUID(id))
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.AdminBadRequestError[*account.AccountJoined](err)
	}

	return response.Success(accountObj)
}

//	 Creates a  Account entry
//
//		@Title			Create Account
//
//		@Summary		Create Account
//		@Description	Create Account
//		@Tags			Account
//		@Accept			json
//		@Produce		json
//		@Param			data	body		account.Account	true	"Account Data"
//		@Success		200	{object}	response.SuccessResponse{data=account.Account}
//		@Failure		400	{object}	response.ErrorResponse
//		@Failure		404	{object}	response.ErrorResponse
//		@Failure		500	{object}	response.ErrorResponse
//		@Router			/admin/account/ [post]
func adminCreate(_ http.ResponseWriter, req *http.Request) (*account.Account, int, error) {
	userSession := request.GetReqSession(req)
	data := request.GetModelPostData(req)
	accountObj := account.New()
	accountObj.MergeData(data)
	err := accountObj.Save(userSession.User)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.AdminBadRequestError[*account.Account](err)

	}

	return response.Success(accountObj)
}

//	 Updates a  Account entry
//
//		@Title			Update Account
//
//		@Summary		Update Account
//		@Description	Update Account
//		@Tags			Account
//		@Accept			json
//		@Produce		json
//	 @Param 			id	path		string	true	"Account ID"
//		@Param			data	body		account.Account	true	"Account Data"
//		@Success		200	{object}	response.SuccessResponse{data=account.Account}
//		@Failure		400	{object}	response.ErrorResponse
//		@Failure		404	{object}	response.ErrorResponse
//		@Failure		500	{object}	response.ErrorResponse
//		@Router			/admin/account/{id} [put]
func adminUpdate(_ http.ResponseWriter, req *http.Request) (*account.AccountJoined, int, error) {
	userSession := request.GetReqSession(req)
	data := request.GetModelPostData(req)
	id := chi.URLParam(req, "id")
	accountObj, err := account.GetJoined(req.Context(), types.UUID(id))
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.AdminBadRequestError[*account.AccountJoined](err)
	}

	if tools.Empty(accountObj) {
		return response.AdminBadRequestError[*account.AccountJoined](errors.Errorf("Object not found with ID: %s", id))
	}

	accountObj.MergeData(data)
	err = accountObj.Save(userSession.User)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.AdminBadRequestError[*account.AccountJoined](err)
	}

	return response.Success(accountObj)
}

//	 Gets the count of Account entries
//
//		@Title			Count Account
//
//		@Summary		Count Account
//		@Description	Count Account
//		@Tags			Account
//		@Accept			json
//		@Produce		json
//		@Param			q	query		string	false	"search by q"
//		@Param			filters	query		string	false	"filters, see readme"
//		@Success		200	{object}	response.SuccessResponse{data=int64}
//		@Failure		400	{object}	response.ErrorResponse
//		@Failure		404	{object}	response.ErrorResponse
//		@Failure		500	{object}	response.ErrorResponse
//		@Router			/admin/account/count [get]
func adminCount(_ http.ResponseWriter, req *http.Request) (int64, int, error) {
	parameters := request.BuildIndexParams(req.Context(), req.URL.Query(), TABLE_NAME)
	account.AddJoinData(parameters)
	count, err := account.FindResultsCount(req.Context(), parameters)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.AdminBadRequestError[int64](err)
	}

	return response.Success(count)
}