# internal/controllers/accounts/x_gen_auth.go
// Code generated by core_gen; DO NOT EDIT.

package accounts

import (
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/griffnb/core/lib/log"
	"github.com/griffnb/core/lib/router/request"
	"github.com/griffnb/core/lib/router/response"
	"github.com/griffnb/core/lib/tools"
	"github.com/griffnb/core/lib/types"
	"github.com/griffnb/techboss-ai-go/internal/constants"
	"github.com/griffnb/techboss-ai-go/internal/models/account"
)

// Gets a list of Account entries
//
//	@Title			List Account
//	@Public
//	@Summary		List Account
//	@Description	List Account
//	@Tags			Account
//	@Accept			json
//	@Produce		json
//	@Param			q	query		string	false	"search by q"
//	@Param			limit	query		int		false	"limit results"		default(100)	minimum(1)	maximum(1000)
//	@Param			offset	query		int		false	"offset results"	default(0)	minimum(0)
//	@Param			order	query		string	false	"sort results e.g. 'created_at desc'"	default(created_at desc)
//	@Param			filters	query		string	false	"filters, see readme"
//	@Success		200	{object}	response.SuccessResponse{data=[]account.AccountJoined}
//	@Failure		400	{object}	response.ErrorResponse
//	@Failure		404	{object}	response.ErrorResponse
//	@Failure		500	{object}	response.ErrorResponse
//	@Router			/account [get]
func authIndex(_ http.ResponseWriter, req *http.Request) ([]*account.AccountJoined, int, error) {

	user := request.GetReqSession(req).User

	parameters := request.BuildIndexParams(req.Context(), req.URL.Query(), TABLE_NAME)

	if tools.Empty(parameters.Limit) {
		parameters.Limit = constants.SYSTEM_LIMIT
	}

	if !tools.Empty(req.URL.Query().Get("q")) {
		addSearch(parameters, req.URL.Query().Get("q"))
	}

	accountObjs, err := account.FindAllRestrictedJoined(req.Context(), parameters, user)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.PublicBadRequestError[[]*account.AccountJoined]()

	}

	return response.Success(accountObjs)
}

//	 Gets a  Account entry by ID
//
//		@Title			Get Account
//		@Public
//		@Summary		Get Account
//		@Description	Get Account
//		@Tags			Account
//		@Accept			json
//		@Produce		json
//		@Param			id	path		string	true	"Account ID"
//		@Success		200	{object}	response.SuccessResponse{data=account.AccountJoined}
//		@Failure		400	{object}	response.ErrorResponse
//		@Failure		404	{object}	response.ErrorResponse
//		@Failure		500	{object}	response.ErrorResponse
//		@Router			/account/{id} [get]
func authGet(_ http.ResponseWriter, req *http.Request) (*account.AccountJoined, int, error) {

	user := request.GetReqSession(req).User

	id := chi.URLParam(req, "id")
	accountObj, err := account.GetRestrictedJoined(req.Context(), types.UUID(id), user)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.PublicBadRequestError[*account.AccountJoined]()

	}

	return response.Success(accountObj)
}

//	 Creates a  Account entry
//
//		@Title			Create Account
//		@Public
//		@Summary		Create Account
//		@Description	Create Account
//		@Tags			Account
//		@Accept			json
//		@Produce		json
//		@Param			data	body		account.Account	true	"Account Data"
//		@Success		200	{object}	response.SuccessResponse{data=account.Account}
//		@Failure		400	{object}	response.ErrorResponse
//		@Failure		404	{object}	response.ErrorResponse
//		@Failure		500	{object}	response.ErrorResponse
//		@Router			/account/ [post]
func authCreate(_ http.ResponseWriter, req *http.Request) (*account.Account, int, error) {

	user := request.GetReqSession(req).User

	data := request.GetModelPostData(req)
	accountObj := account.NewPublic(data, user)
	err := accountObj.Save(user)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.PublicBadRequestError[*account.Account]()

	}

	return response.Success(accountObj)
}

//	 Updates a  Account entry
//
//		@Title			Update Account
//		@Public
//		@Summary		Update Account
//		@Description	Update Account
//		@Tags			Account
//		@Accept			json
//		@Produce		json
//	 @Param 			id	path		string	true	"Account ID"
//		@Param			data	body		account.Account	true	"Account Data"
//		@Success		200	{object}	response.SuccessResponse{data=account.Account}
//		@Failure		400	{object}	response.ErrorResponse
//		@Failure		404	{object}	response.ErrorResponse
//		@Failure		500	{object}	response.ErrorResponse
//		@Router			/account/{id} [put]
func authUpdate(_ http.ResponseWriter, req *http.Request) (*account.AccountJoined, int, error) {

	user := request.GetReqSession(req).User

	id := chi.URLParam(req, "id")
	data := request.GetModelPostData(req)
	accountObj, err := account.GetRestrictedJoined(req.Context(), types.UUID(id), user)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.PublicBadRequestError[*account.AccountJoined]()
	}

	account.UpdatePublic(&accountObj.Account, data, user)
	err = accountObj.Save(user)
	if err != nil {
		log.ErrorContext(err, req.Context())
		return response.PublicBadRequestError[*account.AccountJoined]()
	}

	return response.Success(accountObj)
}