# swag_docs/swagger.yaml
definitions:
  AdminAccountJoined:
    properties:
      created_at:
        type: string
      created_by_name:
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      email_verified_at_ts:
        type: integer
      external_id:
        type: string
      first_name:
        type: string
      hashed_password:
        type: string
      id:
        type: string
      is_super_user_session:
        type: integer
      last_login_ts:
        type: integer
      last_name:
        type: string
      name:
        type: string
      organization_id:
        type: string
      organization_name:
        type: string
      password_updated_at_ts:
        type: integer
      phone:
        type: string
      properties:
        $ref: '#/definitions/account.Properties'
      role:
        type: integer
      signup_properties:
        $ref: '#/definitions/account.SignupProperties'
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        type: string
      updated_by_name:
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    type: object
  account.Account:
    properties:
      ChangeLogs:
        type: boolean
      Class:
        type: string
      Client:
        $ref: '#/definitions/Client'
      CreatedAtColumn:
        type: string
      IDColumn:
        type: string
      IsVersioned:
        type: boolean
      ManualCache:
        type: boolean
      PackageName:
        type: string
      SavingUser:
        $ref: '#/definitions/github.com/griffnb/core/lib/model/coremodel.Model'
      Standardized:
        type: boolean
      Table:
        type: string
      UpdatedAtColumn:
        type: string
      created_at:
        format: date-time
        type: string
      created_by_urn:
        type: string
      data:
        additionalProperties: {}
        type: object
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      email_verified_at_ts:
        type: integer
      external_id:
        type: string
      fieldMap:
        additionalProperties: {}
        type: object
      first_name:
        type: string
      hashed_password:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      isLoaded:
        type: boolean
      isSaved:
        type: boolean
      last_login_ts:
        type: integer
      last_name:
        type: string
      newData:
        additionalProperties: {}
        type: object
      organization_id:
        format: uuid
        type: string
      password_updated_at_ts:
        type: integer
      phone:
        type: string
      preValues:
        additionalProperties: {}
        type: object
      properties:
        $ref: '#/definitions/account.Properties'
      readerSem:
        format: int32
        type: integer
      role:
        type: integer
      saveData:
        additionalProperties: {}
        type: object
      signup_properties:
        $ref: '#/definitions/account.SignupProperties'
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
      v:
        format: int32
        type: integer
      writerSem:
        format: int32
        type: integer
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - properties
    - signup_properties
    - hashed_password
    - password_updated_at_ts
    - email_verified_at_ts
    - last_login_ts
    - is_super_user_session
    type: object
  account.AccountJoined:
    properties:
      ChangeLogs:
        type: boolean
      Class:
        type: string
      Client:
        $ref: '#/definitions/Client'
      CreatedAtColumn:
        type: string
      IDColumn:
        type: string
      IsVersioned:
        type: boolean
      ManualCache:
        type: boolean
      PackageName:
        type: string
      SavingUser:
        $ref: '#/definitions/github.com/griffnb/core/lib/model/coremodel.Model'
      Standardized:
        type: boolean
      Table:
        type: string
      UpdatedAtColumn:
        type: string
      created_at:
        format: date-time
        type: string
      created_by_name:
        type: string
      created_by_urn:
        type: string
      data:
        additionalProperties: {}
        type: object
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      email_verified_at_ts:
        type: integer
      external_id:
        type: string
      fieldMap:
        additionalProperties: {}
        type: object
      first_name:
        type: string
      hashed_password:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      isLoaded:
        type: boolean
      isSaved:
        type: boolean
      last_login_ts:
        type: integer
      last_name:
        type: string
      name:
        type: string
      newData:
        additionalProperties: {}
        type: object
      organization_id:
        format: uuid
        type: string
      organization_name:
        type: string
      password_updated_at_ts:
        type: integer
      phone:
        type: string
      preValues:
        additionalProperties: {}
        type: object
      properties:
        $ref: '#/definitions/account.Properties'
      readerSem:
        format: int32
        type: integer
      role:
        type: integer
      saveData:
        additionalProperties: {}
        type: object
      signup_properties:
        $ref: '#/definitions/account.SignupProperties'
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_name:
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
      v:
        format: int32
        type: integer
      writerSem:
        format: int32
        type: integer
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - properties
    - signup_properties
    - hashed_password
    - password_updated_at_ts
    - email_verified_at_ts
    - last_login_ts
    - is_super_user_session
    - name
    - organization_name
    - created_by_name
    - updated_by_name
    type: object
  account.AccountJoinedPublic:
    properties:
      created_at:
        format: date-time
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      last_name:
        type: string
      organization_id:
        format: uuid
        type: string
      phone:
        type: string
      role:
        type: integer
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - is_super_user_session
    type: object
  account.AccountPublic:
    properties:
      created_at:
        format: date-time
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      last_name:
        type: string
      organization_id:
        format: uuid
        type: string
      phone:
        type: string
      role:
        type: integer
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - is_super_user_session
    type: object
...
  account.Properties:
    properties:
      external_user_info: {}
      invite_key:
        type: string
      invite_ts:
        format: int64
        type: integer
      last_seen:
        format: int64
        type: integer
      verify_email_key:
        type: string
    type: object
...
  account.SignupProperties:
    properties:
      is_oauth:
        format: int64
        type: integer
    required:
    - is_oauth
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /account:
    get:
      consumes:
      - application/json
      description: List Account
      parameters:
      - description: search by q
        in: query
        name: q
        type: string
      - default: 100
        description: limit results
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: offset results
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: created_at desc
        description: sort results e.g. 'created_at desc'
        in: query
        name: order
        type: string
      - description: filters, see readme
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/account.AccountJoinedPublic'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List Account
      tags:
      - Account
  /account/:
    post:
      consumes:
      - application/json
      description: Create Account
      parameters:
      - description: Account Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.AccountPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create Account
      tags:
      - Account
  /account/{id}:
    delete:
      consumes:
      - application/json
      description: Marks an account as deleted by setting status to USER_DELETED
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/AdminAccountJoined'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete account
      tags:
      - Account
    get:
      consumes:
      - application/json
      description: Get Account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountJoinedPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Account
      tags:
      - Account
    put:
      consumes:
      - application/json
      description: Update Account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.AccountPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update Account
      tags:
      - Account
...
  /admin/account:
    get:
      consumes:
      - application/json
      description: List Account
      parameters:
      - description: search by q
        in: query
        name: q
        type: string
      - default: 100
        description: limit results
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: offset results
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: created_at desc
        description: sort results e.g. 'created_at desc'
        in: query
        name: order
        type: string
      - description: filters, see readme
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/account.AccountJoined'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List Account
      tags:
      - Account
  /admin/account/:
    post:
      consumes:
      - application/json
      description: Create Account
      parameters:
      - description: Account Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.Account'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create Account
      tags:
      - Account
  /admin/account/{id}:
    get:
      consumes:
      - application/json
      description: Get Account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountJoined'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Account
      tags:
      - Account
    put:
      consumes:
      - application/json
      description: Update Account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.Account'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update Account
      tags:
      - Account
  /admin/account/count:
    get:
      consumes:
      - application/json
      description: Count Account
      parameters:
      - description: search by q
        in: query
        name: q
        type: string
      - description: filters, see readme
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Count Account
      tags:
      - Account
...
swagger: "2.0"